Move possible alert value to basic_objects
