+++ src/basic_objects.rs
@@ -82,6 +82,17 @@ pub struct LogFile {
     pub file_path: PathBuf,
 }
 
+#[derive(Debug)]
+pub struct PossibleAlertValues {
+    pub num_records: usize,
+    pub num_dupes: usize,
+    pub num_redactions: usize,
+    pub largest_time_gap: Option<TimeGap>,
+    pub errors: Vec<LavaError>,
+    pub mean: f64,
+    pub std: f64,
+}
+
 #[derive(Debug, Default)]
 pub struct ProcessedLogFile {
     pub sha256hash: Option<String>,
