+++ src/processing_objects.rs
@@ -317,14 +317,3 @@ impl LogRecordProcessor {
         self.errors.push(error_to_add);
     }
 }
-
-#[derive(Debug)]
-pub struct PossibleAlertValues {
-    pub num_records: usize,
-    pub num_dupes: usize,
-    pub num_redactions: usize,
-    pub largest_time_gap: Option<TimeGap>,
-    pub errors: Vec<LavaError>,
-    pub mean: f64,
-    pub std: f64,
-}
